backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
azure-storage-blob
werkzeug
